Added the implementation for Signup View as well as the skeleton and high level implementation for the sign interface adapters and use_cases.
I had to restructure the signup process based on our initial design. With this structure, the user first picks their account type, and then provides the inputs. If we ever want to add a user type, we just add the button and action listener in the ChooseAccountTypeView, and add the relevant account type views and business logic.

Each account type has isolated business logic for now in case we need to add account-type specific logic. If later we find that there is nothing new to add, we can simplify it.

I've also added psvm's in each view so you guys can test out what everything looks like.

 I'll start working on creating an InMemory Database and temp Appbuilder to test the signup logic.